reorganize all grunt tasks into one directory
also fixes #12581
